Fix moving the Dragon in PnD expansion
In the Princesses and Dragons expansion, there is a bug that arose from the lua refactoring that was done in this fork.

Currently, upon drawing a dragon tile with the dragon on the board, the dragon movement sequence starts correctly.
However, the dragon movement buttons throw script errors, because the functions tied to the buttons are inaccessible due to being `local`.  Removing the `local` from the four functions fixes this issue and allows the dragon movement to proceed properly.